+++ src/aspire/utils/rotation.py
@@ -256,7 +256,7 @@ class Rotation:
         """
         dtype = dtype or getattr(values, "dtype", np.float32)
         rotations = sp_rot.from_euler("ZYZ", values, degrees=False)
-        matrices = rotations.as_matrix().astype(dtype, copy=False)
+        matrices = rotations.as_matrix().astype(dtype, copy=None)
         return Rotation(matrices)
 
     @staticmethod
@@ -324,7 +324,7 @@ class Rotation:
         """
         dtype = dtype or vec.dtype
         rots = sp_rot.from_rotvec(vec)
-        matrices = rots.as_matrix().astype(dtype, copy=False)
+        matrices = rots.as_matrix().astype(dtype, copy=None)
         return Rotation(matrices)
 
     @staticmethod
@@ -337,7 +337,7 @@ class Rotation:
         :return: new Rotation object
         """
         dtype = dtype or values.dtype
-        return Rotation(values.astype(dtype, copy=False))
+        return Rotation(values.astype(dtype, copy=None))
 
     @staticmethod
     def generate_random_rotations(
