+++ tests/test_coef.py
@@ -90,7 +90,7 @@ def coef_fixture(basis, stack, dtype):
     # shape.
     size = stack + (basis.count,)
 
-    coef_np = np.random.random(size=size).astype(dtype, copy=False)
+    coef_np = np.random.random(size=size).astype(dtype, copy=None)
 
     return Coef(basis, coef_np, dtype=dtype)
 
@@ -212,7 +212,7 @@ def test_add(basis, coef_fixture):
     Tests addition operation against pure Numpy.
     """
     # Make array
-    x = np.random.random(size=coef_fixture.shape).astype(coef_fixture.dtype, copy=False)
+    x = np.random.random(size=coef_fixture.shape).astype(coef_fixture.dtype, copy=None)
     # Construct Coef
     c = Coef(basis, x)
 
@@ -231,7 +231,7 @@ def test_sub(basis, coef_fixture):
     Tests subtraction operation against pure Numpy.
     """
     # Make array
-    x = np.random.random(size=coef_fixture.shape).astype(coef_fixture.dtype, copy=False)
+    x = np.random.random(size=coef_fixture.shape).astype(coef_fixture.dtype, copy=None)
     # Construct Coef
     c = Coef(basis, x)
 
@@ -264,7 +264,7 @@ def test_mul(basis, coef_fixture):
     Tests multiplication operation against pure Numpy.
     """
     # Make array
-    x = np.random.random(size=coef_fixture.shape).astype(coef_fixture.dtype, copy=False)
+    x = np.random.random(size=coef_fixture.shape).astype(coef_fixture.dtype, copy=None)
     # Construct Coef
     c = Coef(basis, x)
